requirements.txt
Flask==3.0.0
google-generativeai
python-dotenv
requests
gunicorn